--- app/index.start.html
@@ -1,5 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-	<meta http-equiv="refresh" content="0;URL='language.html?mode=redirect'" />
-</head>
