+++ app/package.json
@@ -97,8 +97,7 @@
   },
   "devDependencies": {
     "electron": "^4.0.6",
-    "electron-packager": "^13.0.1",
-    "electron-prebuilt": "^1.4.13",
+    "electron-packager": "^13.1.1",
     "electron-rebuild": "^1.8.4"
   }
 }
