+++ app/view/make/index.sample.html
@@ -360,7 +360,6 @@ var tlid=0;
 </div>
 	<div class="modal-footer">
 	<a href="#!" class="waves-effect waves-green btn-flat dm-hide" onclick="brws()">{{openBrowser}}</a>
-	<a href="#!" class="waves-effect waves-green btn-flat dm-hide" onclick="shot()">{{screenshot}}</a>
 	<a href="#!" class="waves-effect waves-green btn-flat dm-hide" onclick="cbCopy()">{{copyURL}}</a>
 	<a href="#!" class="waves-effect waves-green btn-flat dm-hide" onclick="cbCopy('emb')">{{embed}}</a>
 	<a href="#!" class="modal-action modal-close waves-effect waves-green btn-flat">{{close}}</a>
